Job Module added sepratly
1. /duty command added
2. setDuty function added
3. Seprate module added for job.
