+++ resources/urp-bennys/client/main.js
@@ -68,7 +68,7 @@ const isSourceAtBennys = () => {
 }
 
 const isSourceAMechanic = () => {
-   const currentJob = Core.Functions.getJobInfo('name').toLowerCase()
+   const currentJob = Core.Functions.getJobInfo('name')
    const onDuty = Core.Functions.getJobInfo('onDuty')
    if(currentJob !== 'mechanic') {
        alt.emit('notify',  'error', `Benny's`, `You need to be a mechanic`)
