+++ resources/urp-core/client/modules/functions.js
@@ -34,12 +34,8 @@ const getPlayerData = (key) => {
 }
 
 const getJobInfo = (key) => {
-    if(!alt.Player.local.playerData.job) return undefined;
-    if(!alt.Player.local.playerData.job[key]) {
-        return undefined;
-    } else {
+    if(!alt.Player.local.playerData.job[key]) return undefined;
         return alt.Player.local.playerData.job[key];
-    }
 }
 
 const handleDeath = (value) => {
